+++ src/Components/App/App.tsx
@@ -1,7 +1,7 @@
 import React from "react"
-import logo from "../logo.svg"
+import logo from "./logo.svg"
 import "./App.css"
-import { NavBar } from "./NavBar"
+import { NavBar } from "../NavBar/NavBar"
 
 function App() {
 	return (
